Added feature to allow clicking on an entity so that the displayed entities stays.
Changed the module name to include those new features and updated the documentation.
