+++ engine/community-modules/interactivedisplay/Authors.md
@@ -1,3 +1,3 @@
-# CameraModule
+# Interactive display module
 
 Contributed by [Butanium](https://github.com/Butanium).
