+++ engine/community-modules/interactivedisplay/pom.xml
@@ -10,10 +10,10 @@
 	</parent>
 
     <groupId>com.codingame.gameengine</groupId>
-    <artifactId>module-displayonhover</artifactId>
-	<name>CodinGame Game Engine Display-on-hover Module</name>
-	<description>The DisplayOnHoverModule allows you to display entities when the mouse is over an entity. It's the
-		same as the TooltipModule, but with custom entities instead of text.</description>
+    <artifactId>module-interactivedisplay</artifactId>
+	<name>CodinGame Game Engine Interactive display module</name>
+	<description>The InteractiveDisplayModule allows you to display entities when the mouse is over an entity or
+		when the user clicks an entity.</description>
 
 	<dependencies>
 		<dependency>
